💩 Tried to fix the RepetitionParser type definition by adding the MinToMaxTuple type
The biggest achievement of this commit is the development of the MinToMaxTuple generic type.
Due to the MinToMaxTuple type, return type problems for exp.zeroOrMore and exp.oneOrMore have been fixed.
However, the structure of the MinToMaxTuple type is complicated and cannot prevent compilation errors.

We need to look for other solutions.
The RepetitionParser class is used only in exp.zeroOrMore, exp.oneOrMore and exp.times().
Therefore, highly general-purpose types such as MinToMaxTuple are not required.